third_party/blink/web_tests/external/wpt/html/semantics/embedded-content/the-img-element/source-media-outside-doc-expected.txt

This is a testharness.js-based test.
FAIL Image source selection using media queries is performed for img elements outside the document assert_equals: After removing img expected "data:,c" but got "data:,b"
Harness: the test ran to completion.
